pyproject.toml
[tool.poetry]
name = "python-microservice-template"
version = "0.1.0"
description = ""
authors = ["DileepEverestek <100899517+DileepEverestek@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.17.6"
fastapi = "^0.75.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^1.9.0"}
SQLAlchemy = "^1.4.32"
python-dotenv = "^0.20.0"
PyMySQL = "^1.0.2"
python-multipart = "^0.0.5"
cryptography = "^36.0.2"
python-decouple = "^3.6"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
